// src/services/api/generated/models/os-version-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `ubuntu_18_04` - Ubuntu 18.04 LTS * `ubuntu_20_04` - Ubuntu 20.04 LTS * `ubuntu_22_04` - Ubuntu 22.04 LTS * `ubuntu_24_04` - Ubuntu 24.04 LTS * `centos_7` - CentOS 7 * `centos_8` - CentOS 8 * `centos_9` - CentOS 9 * `windows_10` - Windows 10 * `windows_11` - Windows 11 * `macos_ventura` - macOS Ventura * `macos_sonoma` - macOS Sonoma
 */

export const OsVersionEnum = {
    ubuntu_18_04: 'ubuntu_18_04',
    ubuntu_20_04: 'ubuntu_20_04',
    ubuntu_22_04: 'ubuntu_22_04',
    ubuntu_24_04: 'ubuntu_24_04',
    centos_7: 'centos_7',
    centos_8: 'centos_8',
    centos_9: 'centos_9',
    windows_10: 'windows_10',
    windows_11: 'windows_11',
    macos_ventura: 'macos_ventura',
    macos_sonoma: 'macos_sonoma'
} as const;

export type OsVersionEnum = typeof OsVersionEnum[keyof typeof OsVersionEnum];